requirements.txt
numpy==2.3.1
cvxpy==1.7.1
PyYAML==6.0.2
pandas==2.3.1